requirements/base.txt
# Core
Flask>=2.3
Flask-SQLAlchemy>=3.0
Flask-Migrate>=4.0
Flask-JWT-Extended>=4.6
Flask-Cors>=4.0
python-dotenv>=1.0

# Banco de Dados
psycopg2-binary>=2.9

# Segurança
Werkzeug>=2.3
passlib>=1.7

# Agendador
APScheduler>=3.10

# Templates & Relatórios
Jinja2>=3.1
WeasyPrint>=61.0  # Para geração de PDF

# Validação
marshmallow>=3.19

# Utilitários
requests>=2.31
